improved makefile

+++ src/string.cpp
@@ -12,12 +12,11 @@ char* itoa(uint32_t val) {
 	if (val == 0)
 		return "0";
 	static char buf[32] = {0};
-	const uint64_t BASE = 10;
-	
-	uint64_t i = 30;
+	const uint8_t BASE = 10;
+
+	uint32_t i = 30;
 	
 	for(; val && i ; --i, val /= BASE)
-	
 		buf[i] = "0123456789abcdef"[val % BASE];
 	
 	return &buf[i+1];
